refactor: split up mint panels
