+++ libs/admin/app/ui/src/index.ts
@@ -18,3 +18,5 @@ export * from './lib/admin-app-ui-wallet-balances'
 export * from './lib/admin-app-ui-webhook-detail'
 export * from './lib/admin-app-ui-webhooks'
 export * from './lib/show-sol-balance'
+export { AdminAppUiMintEnabledPanel } from './lib/admin-app-ui-mint-enabled-panel'
+export { AdminAppUiMintDisabledPanel } from './lib/admin-app-ui-mint-disabled-panel'
