+++ libs/admin/app/ui/src/lib/admin-app-ui-app-env-mint-settings.tsx
@@ -1,9 +1,9 @@
-import { SettingsIcon } from '@chakra-ui/icons'
-import { Box, Button, Code, Flex, SimpleGrid, Stack, Text } from '@chakra-ui/react'
-import { AdminClusterUiMintDetails } from '@mogami/admin/cluster/ui'
+import { Box, SimpleGrid, Stack, Text } from '@chakra-ui/react'
 import { AdminUiAlert } from '@mogami/admin/ui/alert'
-import { AppEnv, Mint, Wallet } from '@mogami/shared/util/admin-sdk'
-import React, { useState } from 'react'
+import { AppEnv } from '@mogami/shared/util/admin-sdk'
+import React from 'react'
+import { AdminAppUiMintDisabledPanel } from './admin-app-ui-mint-disabled-panel'
+import { AdminAppUiMintEnabledPanel } from './admin-app-ui-mint-enabled-panel'
 
 export interface AdminAppUiAppEnvMintSettingsProps {
   appEnv: AppEnv
@@ -30,7 +30,7 @@ export function AdminAppUiAppEnvMintSettings({
       <Text>Enabled Mints</Text>
       <SimpleGrid columns={2} gap={6}>
         {appEnv?.mints?.map((mint) => (
-          <EnabledMintPanel
+          <AdminAppUiMintEnabledPanel
             key={mint?.id}
             mint={mint!.mint}
             wallet={mint!.wallet}
@@ -45,7 +45,7 @@ export function AdminAppUiAppEnvMintSettings({
       {clusterMints?.length ? (
         <SimpleGrid columns={2} gap={6}>
           {clusterMints?.map((mint) => (
-            <DisabledMintPanel key={mint?.id} mint={mint} enableMint={enableMint} />
+            <AdminAppUiMintDisabledPanel key={mint?.id} mint={mint} enableMint={enableMint} />
           ))}
         </SimpleGrid>
       ) : (
@@ -53,63 +53,4 @@ export function AdminAppUiAppEnvMintSettings({
       )}
     </Stack>
   )
-}
-
-function DisabledMintPanel({ enableMint, mint }: { enableMint: (mintId: string) => void; mint?: Mint | null }) {
-  return mint && mint?.id ? (
-    <Stack direction="column" spacing={6} p={6} borderWidth="1px" borderRadius="lg" key={mint?.id}>
-      {mint && <AdminClusterUiMintDetails mint={mint} />}
-      <Flex justifyContent="space-between" alignItems="center">
-        <Button size="lg" onClick={() => enableMint(mint!.id!)}>
-          Enable {mint?.symbol}
-        </Button>
-      </Flex>
-    </Stack>
-  ) : null
-}
-
-function EnabledMintPanel({
-  disableMint,
-  mint,
-  selectWallet,
-  wallet,
-  wallets,
-}: {
-  disableMint: (mintId: string) => void
-  mint?: Mint | null
-  selectWallet: (walletId: string) => void
-  wallet?: Wallet | null
-  wallets?: Wallet[]
-}) {
-  const [showDetails, setShowDetails] = useState(false)
-  const toggleDetails = () => setShowDetails((value) => !value)
-  const availableWallets = wallets?.filter((item) => item.id !== wallet?.id)
-  return mint && mint?.id ? (
-    <Stack direction="column" spacing={6} p={6} borderWidth="1px" borderRadius="lg">
-      <AdminClusterUiMintDetails mint={mint} />
-      <Flex justifyContent="space-between" alignItems="center">
-        <Button size="lg" onClick={toggleDetails}>
-          <SettingsIcon /> <Text ml={2}>Settings</Text>
-        </Button>
-      </Flex>
-      {showDetails ? (
-        <Stack direction="column" spacing={6}>
-          <Code colorScheme="teal">{wallet?.publicKey}</Code>
-          {availableWallets?.map((item) => (
-            <Flex justifyContent="space-between" alignItems="center" key={item.id}>
-              <Code colorScheme="teal">{item?.publicKey}</Code>
-              <Button size="sm" onClick={() => selectWallet(item.id!)}>
-                Use as Fee Payer
-              </Button>
-            </Flex>
-          ))}
-          <Flex justifyContent="space-between" alignItems="center">
-            <Button size="lg" disabled={!!mint?.default} onClick={() => disableMint(mint.id!)}>
-              {mint?.default ? 'Default mint' : `Disable ${mint?.symbol}`}
-            </Button>
-          </Flex>
-        </Stack>
-      ) : null}
-    </Stack>
-  ) : null
 }
